// apps/main/src/models/v2/controlplane/moduleProviderSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */

/**
 * A summary of configuration for an OpenTofu provider that will be used in modules
 */
export interface ModuleProviderSummary {
  /** The Organization ID */
  org_id: string;
  /** The OpenTofu provider name */
  provider_type: string;
  /** The unique identifier for this module provider under the given provider type */
  id: string;
  /** An optional text description for this module provider */
  description?: string;
  /** The OpenTofu provider source as a registry coordinate */
  source: string;
  /** The date and time when the provider was created. */
  created_at: string;
}